+++ CHANGELOG.md
@@ -13,6 +13,10 @@
 **Fixed**
 - Fix an issue with updating the "active" row after archiving events.
 
+**Changes**
+- In request to the backend, rename maxTs, minTs, eventType to max_ts,
+  min_ts and event_type.
+
 [Full Changelog](https://github.com/jasonish/evebox/compare/0.8.0...HEAD)
 
 ## 0.8.0 - 2017-06-30